[rt.cpan.org 76550] Avoid hang when no requests made.
Fixed thanks to Sergei Kozunov's bug report and test case.  A sidecar
process was created at startup, but no idle timeout was set.  These
timeouts are only set when requests happen, and in an application
where all addresses are already resolved, no timeout is set.  So the
component lingers forever.